split useReportState into multiple files - refactored some of actions.ts to be more concise
