+++ next-client/src/components/report/hooks/useReportState/__tests__/useReportState.open.test.ts
@@ -1,15 +1,9 @@
 import { describe, test, expect } from "vitest";
-import { ReportState, SomeNode, __internals } from "../useReportState";
-import { reportData } from "stories/data/dummyData";
+import { setupTestState } from "./testStateSetup";
+import { ReportState, SomeNode } from "../types";
+import { defaultSubtopicPagination, defaultTopicPagination } from "../consts";
 
-const {
-  createPathMapReducer,
-  stateBuilder,
-  defaultTopicPagination,
-  mapIdsToPath,
-} = __internals;
-
-const state = stateBuilder(reportData.topics);
+const { state, reducer } = setupTestState();
 
 // Gets a topic at a specific index
 const getTopic = (state: ReportState, idx: number = 0) => state.children[idx];
@@ -35,7 +29,6 @@ const firstTopic = getTopic(state);
 const lastTopic = getLastTopic(state);
 const middleTopic = getMiddleTopic(state);
 const testTopics = [firstTopic, middleTopic, lastTopic];
-const reducer = createPathMapReducer(mapIdsToPath(state));
 const open = (reportState: ReportState, id: string) =>
   reducer(reportState, { type: "open", payload: { id } });
 
@@ -122,7 +115,7 @@ describe("Open", () => {
       });
 
       test("Low subtopic testing idx is smaller than the default topic pagination", () => {
-        expect(testSubtopicLowIdx).lessThan(defaultTopicPagination);
+        expect(testSubtopicLowIdx).lessThan(defaultSubtopicPagination);
       });
     });
     test("Opening to subtopic sets topic isOpen to true", () => {
