+++ next-client/src/components/report/hooks/useReportState/consts.ts
@@ -1,0 +1,10 @@
+//  ********************************
+//  * CONSTS
+//  ********************************/
+
+// Pagination is 0 indexed, so 0 will show one
+export const defaultAddTopicPagination = 1; // 2 additional subtopics shown after each pag
+export const defaultTopicPagination = 1; // 2 subtopics show initially
+
+export const defaultSubtopicPagination = 7; // 6 claims shown initially
+export const defaultAddSubtopicPagination = 9; // 8 additional claims shown after each pag
